+++ customize_local_planner/customize_local_planner/pwm.py
@@ -8,7 +8,7 @@ Created: 10/2/24
 # ROS MODULES
 from rclpy.node import Publisher
 from rclpy.impl.rcutils_logger import RcutilsLogger
-from std_msgs.msg import UInt32, Float64
+from std_msgs.msg import UInt32, Float32
 
 
 class PWM:
@@ -42,7 +42,7 @@ class PWM:
                         self._value = round(value)
                 # publish new value
                 self.value_pub.publish(UInt32(data=self._value))
-                self.percent_pub.publish(Float64(data=self.percentage))
+                self.percent_pub.publish(Float32(data=self.percentage))
         
         @property
         def percentage(self) -> float:
